feat: extract some traits
