+++ crates/storage/provider/Cargo.toml
@@ -20,6 +20,7 @@ reth-primitives.workspace = true
 reth-fs-util.workspace = true
 reth-storage-errors.workspace = true
 reth-interfaces.workspace = true
+reth-storage-api.workspace = true
 reth-db.workspace = true
 reth-trie = { workspace = true, features = ["metrics"] }
 reth-nippy-jar.workspace = true
