+++ crates/storage/provider/src/traits/mod.rs
@@ -1,53 +1,22 @@
 //! Collection of common provider traits.
 
-mod account;
-pub use account::{AccountExtReader, AccountReader, ChangeSetReader};
-
-mod storage;
-pub use storage::StorageReader;
-
-mod block;
-pub use block::{
-    BlockExecutionWriter, BlockReader, BlockReaderIdExt, BlockSource, BlockWriter,
-    TransactionVariant,
-};
-
-mod block_hash;
-pub use block_hash::BlockHashReader;
-
-mod block_id;
-pub use block_id::{BlockIdReader, BlockNumReader};
+// Re-export all the traits
+pub use reth_storage_api::*;
 
 // Re-export for convenience
 pub use reth_evm::provider::EvmEnvProvider;
 
+mod block;
+pub use block::*;
+
 mod chain_info;
 pub use chain_info::CanonChainTracker;
-
-mod header;
-pub use header::HeaderProvider;
 
 mod header_sync_gap;
 pub use header_sync_gap::{HeaderSyncGap, HeaderSyncGapProvider, HeaderSyncMode};
 
-mod receipts;
-pub use receipts::{ReceiptProvider, ReceiptProviderIdExt};
-
 mod state;
-pub use state::{
-    BlockchainTreePendingStateProvider, BundleStateDataProvider, BundleStateForkProvider,
-    FullBundleStateDataProvider, StateProvider, StateProviderBox, StateProviderFactory,
-    StateWriter,
-};
-
-mod trie;
-pub use trie::StateRootProvider;
-
-mod transactions;
-pub use transactions::{TransactionsProvider, TransactionsProviderExt};
-
-mod withdrawals;
-pub use withdrawals::WithdrawalsProvider;
+pub use state::StateWriter;
 
 mod chain;
 pub use chain::{
