+++ crates/storage/storage-api/src/block_hash.rs
@@ -1,9 +1,8 @@
-use auto_impl::auto_impl;
 use reth_primitives::{BlockHashOrNumber, BlockNumber, B256};
 use reth_storage_errors::provider::ProviderResult;
 
 /// Client trait for fetching block hashes by number.
-#[auto_impl(&, Arc, Box)]
+#[auto_impl::auto_impl(&, Arc, Box)]
 pub trait BlockHashReader: Send + Sync {
     /// Get the hash of the block with the given number. Returns `None` if no block with this number
     /// exists.
