+++ crates/storage/storage-api/src/header.rs
@@ -1,10 +1,9 @@
-use auto_impl::auto_impl;
 use reth_primitives::{BlockHash, BlockHashOrNumber, BlockNumber, Header, SealedHeader, U256};
 use reth_storage_errors::provider::ProviderResult;
 use std::ops::RangeBounds;
 
 /// Client trait for fetching `Header` related data.
-#[auto_impl(&, Arc)]
+#[auto_impl::auto_impl(&, Arc)]
 pub trait HeaderProvider: Send + Sync {
     /// Check if block is known
     fn is_known(&self, block_hash: &BlockHash) -> ProviderResult<bool> {
