+++ crates/storage/storage-api/src/lib.rs
@@ -1,0 +1,42 @@
+//! Collection of traits and types for common storage access.
+
+#![doc(
+    html_logo_url = "https://raw.githubusercontent.com/paradigmxyz/reth/main/assets/reth-docs.png",
+    html_favicon_url = "https://avatars0.githubusercontent.com/u/97369466?s=256",
+    issue_tracker_base_url = "https://github.com/paradigmxyz/reth/issues/"
+)]
+#![cfg_attr(not(test), warn(unused_crate_dependencies))]
+#![cfg_attr(docsrs, feature(doc_cfg, doc_auto_cfg))]
+
+mod account;
+pub use account::*;
+
+mod block;
+pub use block::*;
+
+mod block_id;
+pub use block_id::*;
+
+mod block_hash;
+pub use block_hash::*;
+
+mod header;
+pub use header::*;
+
+mod receipts;
+pub use receipts::*;
+
+mod state;
+pub use state::*;
+
+mod storage;
+pub use storage::*;
+
+mod transactions;
+pub use transactions::*;
+
+mod trie;
+pub use trie::*;
+
+mod withdrawals;
+pub use withdrawals::*;
