+++ crates/storage/storage-api/src/transactions.rs
@@ -1,10 +1,11 @@
+use std::ops::{Range, RangeBounds, RangeInclusive};
+
 use crate::{BlockNumReader, BlockReader};
 use reth_primitives::{
     Address, BlockHashOrNumber, BlockNumber, TransactionMeta, TransactionSigned,
     TransactionSignedNoHash, TxHash, TxNumber,
 };
 use reth_storage_errors::provider::{ProviderError, ProviderResult};
-use std::ops::{Range, RangeBounds, RangeInclusive};
 
 ///  Client trait for fetching [TransactionSigned] related data.
 #[auto_impl::auto_impl(&, Arc)]
