+++ crates/storage/storage-api/src/trie.rs
@@ -1,11 +1,10 @@
-use auto_impl::auto_impl;
 use reth_primitives::B256;
 use reth_storage_errors::provider::ProviderResult;
 use reth_trie::updates::TrieUpdates;
 use revm::db::BundleState;
 
 /// A type that can compute the state root of a given post state.
-#[auto_impl(&, Box, Arc)]
+#[auto_impl::auto_impl(&, Box, Arc)]
 pub trait StateRootProvider: Send + Sync {
     /// Returns the state root of the `BundleState` on top of the current state.
     ///
